requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pypdf2>=3.0.0
python-docx>=0.8.11
chromadb>=0.4.0
nltk>=3.6.3
requests>=2.26.0
pydantic-settings>=2.0.0
rich>=10.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.2
sentence-transformers>=2.3.0

# Note: PyTorch needs to be installed separately based on your system:
# For Apple Silicon (M1/M2/M3): pip3 install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cpu
# For Intel Mac: pip3 install torch torchvision torchaudio

